Move java path to evosuite settings page
