+++ src/main/kotlin/org/jetbrains/research/testgenie/settings/SettingsProjectState.kt
@@ -6,7 +6,6 @@ import org.jetbrains.research.testgenie.TestGenieDefaultsBundle
  * This class is the actual data class that stores the values of the Plugin Settings entries.
  */
 data class SettingsProjectState(
-    var javaPath: String = DefaultSettingsPluginState.javaPath,
     var colorRed: Int = DefaultSettingsPluginState.colorRed,
     var colorGreen: Int = DefaultSettingsPluginState.colorGreen,
     var colorBlue: Int = DefaultSettingsPluginState.colorBlue,
