+++ src/main/resources/defaults/Labels.properties
@@ -5,6 +5,8 @@ parametersDepth=The depth of input parameters used in class under tests
 maximumNumberOfRequests=Maximum number of requests to LLM
 maximumPolyDepth=Maximum polymorphism depth
 !EvoSuite Settings
+javaSettings=Java settings
+javaPath=Java 11 path:
 defaultSearch=Select search algorithm
 seed=Seed (random if left empty)
 configId=Select configuration id (null if left empty)
@@ -28,8 +30,7 @@ telemetry=Telemetry
 telemetryEnabled=Enable telemetry
 telemetryPath=Specify the folder path for telemetry data
 accessibility=Accessibility settings
-environmentSettings=Environment settings
-javaPath=Java 11 path:
+environmentSettings=Project compilation settings
 buildPath=Put the compilation path:
 buildCommand=Put the compile command:
 colorPicker=Choose color for visualisation highlight
