+++ src/main/resources/defaults/Tooltips.properties
@@ -14,7 +14,7 @@ buildPath=Directs EvoSuite to the compiled classes. Default: empty string, in th
 buildCommand=The command you use for compiling. Default: empty string, in this case, build goes through the IDE.
 telemetryEnabled=Send telemetry to CISELab
 telemetryPath=Choose a directory to save telemetry data into
-javaPath=Path to a java binary
+javaPath=Path to a java binary for running EvoSuite
 !QuickAccess
 default=Default:
 seconds=seconds
