+++ src/main/resources/messages/TestGenie.properties
@@ -45,11 +45,11 @@ unknownClassError=Unknown class
 errorWhileInitializingTargetClass=Error while initializing target class
 !LLM error messages
 missingToken=The token for the Grazie Platform is not provided. You can set it in the Large Language Model Settings
-emptyResponse=LLM could not generate any tests for this class. Asking the assistant to fix its mistake.
+emptyResponse=LLM could not generate any tests for this class. Asking the AI assistance to fix its mistake.
 emptyBuildPath=Build path is Empty!\nPlease make sure that IDEA recognizes all of your module or enter proper build path in settings
 badGrazieResponse=Problems accessing the Chat GPT server
 invalidGrazieResult=The result is invalid or uses unknown commands due to randomness and lack of guarantees from Chat GPT.\nPlease try again
-compilationError=The test generated by LLM is not compilable. Asking the assistant to fix its mistake.
+compilationError=The test generated by LLM is not compilable. Asking the AI assistance to fix its mistake.
 incorrectToken=The token for the Grazie Platform is not correct
 serverProblems=Grazie Platform is not responding
 tooLongPrompt=The generated prompt is too long!\nPlease reduce maximum polymorphism depth and/or maximum input parameters depth in LLm settings
